+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
  - `i2s-audio-out-dma.rs` example, too difficult to fix.
+- Add `SysMonoTimerExt` helper trait [#497]
+
+[#497]: https://github.com/stm32-rs/stm32f4xx-hal/pull/497
 
 ## [v0.13.2] - 2022-05-16